requirements.txt
dj-static==0.0.6
Django==1.11.4
django-registration==2.2
django-registration-redux==1.7
django-tools==0.33.0
djangorestframework==3.6.4
gunicorn==19.7.1
lxml==3.8.0
olefile==0.44
Pillow==4.2.1
psycopg2==2.7.3
pytz==2017.2
static3==0.7.0

channels==1.1.8
asgi_redis==1.4.3